+++ packages/widgetbook/lib/src/fields/color_field/color_space.dart
@@ -1,0 +1,5 @@
+enum ColorSpace {
+  hex,
+  rgb,
+  hsl,
+}