+++ packages/widgetbook/lib/src/fields/color_field/hex_color_picker.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
+import 'opaque_color.dart';
+import 'opaque_color_picker.dart';
+
+class HexColorPicker extends StatelessWidget implements OpaqueColorPicker {
+  const HexColorPicker({
+    super.key,
+    required this.value,
+    required this.onChanged,
+  });
+
+  final OpaqueColor value;
+  final ValueChanged<OpaqueColor> onChanged;
+
+  String get hexValue => '${value.red.toRadixString(16).padLeft(2, '0')}'
+      '${value.green.toRadixString(16).padLeft(2, '0')}'
+      '${value.blue.toRadixString(16).padLeft(2, '0')}';
+
+  @override
+  Widget build(BuildContext context) {
+    return TextFormField(
+      initialValue: hexValue,
+      maxLength: 6,
+      decoration: const InputDecoration(
+        counterText: '',
+        prefixText: '#',
+      ),
+      inputFormatters: [
+        FilteringTextInputFormatter.allow(
+          RegExp(r'[0-9a-fA-F]'),
+          replacementString: hexValue,
+        ),
+      ],
+      onChanged: (value) {
+        final source = int.tryParse(
+          value,
+          radix: 16,
+        );
+
+        if (value.length != 6 || source == null) return;
+        onChanged.call(OpaqueColor(source));
+      },
+    );
+  }
+}