+++ packages/widgetbook/lib/src/knobs/builders/knobs_builder.dart
@@ -64,12 +64,15 @@ class KnobsBuilder {
   Color color({
     required String label,
     required Color initialValue,
+    ColorSpace initialColorSpace = ColorSpace.hex,
     String? description,
   }) {
     return onKnobAdded(
       ColorKnob(
         label: label,
         value: initialValue,
+        initialColorSpace: initialColorSpace,
+        description: description,
       ),
     )!;
   }